requirements.txt
# Stripe Comptable Export - Dependencies
# Python 3.9+ required

# Stripe API SDK
stripe==7.10.0

# Environment variable management
python-dotenv==1.0.0

# Data manipulation and CSV export
pandas==2.1.4

# Excel export
openpyxl==3.1.2

# PDF generation
reportlab==4.0.8

# CLI framework
click==8.1.7

# HTTP requests for downloading invoice PDFs
requests==2.31.0

# Date/time handling
python-dateutil==2.8.2

